In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,14 +7,14 @@
  */
 int main(void)
 {
-	unsigned long int first = 1, sec = 2, next;
+	unsigned int first = 1, sec = 2, next;
 	int i = 0;
 
-	printf("%lu, %lu", first, sec);
+	printf("%u, %u", first, sec);
 	while (i < 98)
 	{
 		next = first + sec;
-		printf(", %lu", next);
+		printf(", %u", next);
 		first = sec;
 		sec = next;
 		i++;
